Project 1/Provisioning/requirements.txt
azure-identity
azure-mgmt-network
azure-functions
requests

